feat(marketing): render bilan offer context server-side
