+++ app/bilan-gratuit/BilanStrategiqueClient.tsx
@@ -1,34 +1,22 @@
 'use client';
 
-import { useEffect, useMemo, useState, type FormEvent } from 'react';
+import { useEffect, useState, type FormEvent } from 'react';
 import Link from 'next/link';
-import { useRouter, useSearchParams } from 'next/navigation';
+import { useRouter } from 'next/navigation';
 import { CheckCircle2, GraduationCap, Phone } from 'lucide-react';
 import { WhatsAppLogo, WHATSAPP_BRAND_GREEN } from '@/components/ui/whatsapp-logo';
 import { toast, Toaster } from 'sonner';
 import { track } from '@/lib/analytics';
 import { CorporateFooter } from '@/components/layout/CorporateFooter';
-import { Badge } from '@/components/ui/badge';
 import { Card, CardContent } from '@/components/ui/card';
 import { Checkbox } from '@/components/ui/checkbox';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { Textarea } from '@/components/ui/textarea';
 import { ConseillerCard, ProcessSteps } from '@/components/marketing/acadomia-inspired';
-import { PaymentMethodsNote } from '@/components/marketing/PaymentMethodsNote';
-import { fmtTND } from '@/components/premium/format';
-import {
-  getAnnualOffer,
-  getAnnualOfferPaymentSchedule,
-  getCarte,
-  getCoachingOffer,
-  getEffectivePrice,
-  getPack,
-  getPonctuelOffer,
-  getStageFormat,
-} from '@/lib/pricing';
 import { buildWhatsAppUrl } from '@/lib/whatsapp';
 import { LEGAL } from '@/lib/legal';
+import type { SelectedOfferContext } from './selected-offer';
 
 const SUBJECTS = [
   { value: 'MATHEMATIQUES', label: 'Mathématiques' },
@@ -77,117 +65,16 @@ const initialFormData: FormData = {
 };
 
 
-/** Canonical payment breakdown — same structure used by OfferDetailDialog. */
-export type SelectedOfferContext = {
-  id: string;
-  title: string;
-  price: number;
-  deposit?: number;
-  /** Individual installment amounts (same values the modal renders). */
-  installments?: number[];
-  /** Single solde amount (stages, ponctuel). */
-  solde?: number;
-  /** Multiple solde amounts (packs, coaching with schedule). */
-  solde_schedule?: number[];
-  /** Full payment at booking (no échéancier). */
-  full_at_booking?: boolean;
+type BilanStrategiqueClientProps = {
+  programme?: string | null;
+  selectedOffer?: SelectedOfferContext | null;
 };
 
-export function resolveSelectedOfferContext(id: string | null): SelectedOfferContext | null {
-  if (!id) return null;
-
-  const annual = getAnnualOffer(id);
-  if (annual) {
-    const price = getEffectivePrice(annual);
-    const payment = getAnnualOfferPaymentSchedule(annual);
-    if (!price) return null;
-    return {
-      id: annual.id,
-      title: annual.title,
-      price,
-      deposit: payment?.deposit,
-      installments: payment?.installments,
-    };
-  }
-
-  const stage = getStageFormat(id);
-  if (stage) {
-    return {
-      id: stage.format_id,
-      title: stage.title,
-      price: stage.price_per_student,
-      deposit: stage.payment.deposit,
-      solde: stage.payment.solde,
-    };
-  }
-
-  const ponctuel = getPonctuelOffer(id);
-  if (ponctuel) {
-    return {
-      id: ponctuel.id,
-      title: ponctuel.title,
-      price: ponctuel.price_per_student,
-      deposit: ponctuel.payment.deposit,
-      solde: ponctuel.payment.full_at_booking ? undefined : ponctuel.payment.solde,
-      full_at_booking: ponctuel.payment.full_at_booking,
-    };
-  }
-
-  const coaching = getCoachingOffer(id);
-  if (coaching) {
-    return {
-      id: coaching.id,
-      title: coaching.title,
-      price: coaching.price,
-      deposit: coaching.payment.full_at_booking ? coaching.price : coaching.payment.deposit,
-      solde: coaching.payment.solde,
-      solde_schedule: coaching.payment.solde_schedule,
-      full_at_booking: coaching.payment.full_at_booking,
-    };
-  }
-
-  const pack = getPack(id);
-  if (pack) {
-    return {
-      id: pack.id,
-      title: pack.title,
-      price: pack.price,
-      deposit: pack.payment.deposit,
-      solde_schedule: pack.payment.solde_schedule,
-    };
-  }
-
-  const carte = getCarte();
-  if (carte.id === id) {
-    return {
-      id: carte.id,
-      title: carte.title,
-      price: carte.price_annual,
-      full_at_booking: true,
-      deposit: carte.price_annual,
-    };
-  }
-
-  return null;
-}
-
-export function BilanStrategiqueClient() {
+export function BilanStrategiqueClient({
+  programme = null,
+  selectedOffer = null,
+}: BilanStrategiqueClientProps) {
   const router = useRouter();
-  const searchParams = useSearchParams();
-  const programme = searchParams?.get('programme');
-  const offerId = searchParams?.get('offer');
-  const selectedOffer = useMemo(() => resolveSelectedOfferContext(offerId), [offerId]);
-  const programmeLabel = useMemo(() => {
-    if (!programme) return null;
-    const labels: Record<string, string> = {
-      excellence: 'Excellence',
-      plateforme: 'Plateforme',
-      hybride: 'Hybride',
-      immersion: 'Immersion',
-      'pack-specialise': 'Pack spécialisé',
-    };
-    return labels[programme] ?? programme;
-  }, [programme]);
 
   const [formData, setFormData] = useState<FormData>(initialFormData);
   const [selectedSubjects, setSelectedSubjects] = useState<string[]>([]);
@@ -277,68 +164,6 @@ export function BilanStrategiqueClient() {
   return (
     <>
       {/* Toaster is provided globally by components/providers.tsx */}
-
-      <section className="bg-lux-ink px-4 py-16 pt-28 md:px-6">
-        <div className="mx-auto max-w-5xl text-center">
-          <Badge className="mb-4 border border-lux-line/40 bg-white/5 text-lux-gold-wash">
-            Diagnostic gratuit
-          </Badge>
-          <h1 className="font-fraunces text-4xl font-light tracking-tight text-lux-ivory md:text-5xl">
-            Bilan stratégique gratuit
-          </h1>
-          <p className="mx-auto mt-4 max-w-2xl text-lg text-lux-on-dark-muted">
-            Identifier les priorités de votre enfant avant de choisir une formule. Réponse personnalisée,
-            orientation vers la bonne solution et échange humain avec notre équipe pédagogique.
-          </p>
-          {programmeLabel && (
-            <p className="mt-3 text-sm text-lux-gold-wash">
-              Contexte repéré&nbsp;: {programmeLabel}
-            </p>
-          )}
-          {selectedOffer && (
-            <div className="mx-auto mt-5 max-w-2xl rounded-2xl border border-lux-line/30 bg-white/5 p-4 text-left text-sm text-lux-on-dark-muted">
-              <p className="font-semibold text-lux-gold-wash">Offre repérée&nbsp;: {selectedOffer.title}</p>
-              <p className="mt-1">
-                Tarif {fmtTND(selectedOffer.price)}
-                {selectedOffer.deposit != null && !selectedOffer.full_at_booking
-                  ? ` · acompte ${fmtTND(selectedOffer.deposit)}`
-                  : ''}
-                {selectedOffer.full_at_booking
-                  ? ` · règlement intégral à la réservation`
-                  : selectedOffer.installments && selectedOffer.installments.length > 0
-                    ? ` · ${selectedOffer.installments.length}\u00A0×\u00A0${fmtTND(selectedOffer.installments[0])}${
-                        selectedOffer.installments[selectedOffer.installments.length - 1] !== selectedOffer.installments[0]
-                          ? ` puis ${fmtTND(selectedOffer.installments[selectedOffer.installments.length - 1])}`
-                          : ''
-                      }`
-                    : selectedOffer.solde != null
-                      ? ` · solde ${fmtTND(selectedOffer.solde)}`
-                      : selectedOffer.solde_schedule && selectedOffer.solde_schedule.length > 0
-                        ? ` · ${selectedOffer.solde_schedule.length} soldes`
-                        : ''}
-              </p>
-              <div className="mt-4">
-                <PaymentMethodsNote tone="dark" />
-              </div>
-            </div>
-          )}
-
-          <div className="mt-8 flex flex-col items-center justify-center gap-3 sm:flex-row">
-            <a
-              href={buildWhatsAppUrl('le bilan gratuit')}
-              target="_blank"
-              rel="noopener noreferrer"
-              className="lux-cta-secondary rounded-lg px-6 py-3.5 text-sm font-semibold text-lux-ivory border-lux-line/40"
-            >
-              <WhatsAppLogo className="mr-2 h-4 w-4" style={{ color: WHATSAPP_BRAND_GREEN }} />
-              Écrire sur WhatsApp
-            </a>
-            <Link href="/offres" className="lux-cta-reserve rounded-lg px-6 py-3.5 text-sm font-semibold">
-              Voir les offres
-            </Link>
-          </div>
-        </div>
-      </section>
 
       <section className="bg-lux-paper px-4 py-14 md:px-6">
         <div className="mx-auto grid max-w-6xl gap-6 lg:grid-cols-[1.15fr_0.85fr]">
