+++ docs/audits/2026-06-24-lot-2-polish.md
@@ -1,0 +1,61 @@
+# Lot 2 - Decisions produit et polish
+
+## Date
+
+2026-06-24
+
+## Contexte
+
+Le plan de finalisation demande trois sujets Lot 2 : arbitrage Carte Nexus, rendu SSR de `/bilan-gratuit`, et harmonisation tutoiement / vouvoiement.
+
+## Problemes observes
+
+- `/bilan-gratuit?offer=...` rendait un fallback HTML avec `BAILOUT_TO_CLIENT_SIDE_RENDERING` parce que le composant client lisait `useSearchParams`.
+- L'offre reperee et la note de paiement carte etaient visibles apres hydratation, mais pas garanties dans le HTML initial.
+- Carte Nexus reste une decision produit : `carte-nexus` a 290 TND / an inclut `ARIA Autonomie`, vendue en standalone a 590 TND / an.
+
+## Decisions prises
+
+- Deplacer la resolution `programme` / `offer` dans le server component `app/bilan-gratuit/page.tsx`.
+- Sortir le resolver d'offre dans `app/bilan-gratuit/selected-offer.ts`, module partage et testable.
+- Rendre le hero, l'offre reperee et la politique ClicToPay dans le HTML initial.
+- Conserver le tutoiement dans les surfaces pedagogiques eleve (`dashboard/eleve`, ARIA widget, LAMIS, automatismes), car elles s'adressent directement a l'eleve. Les surfaces marketing scannees restent au vouvoiement.
+
+## Fichiers modifies
+
+- `app/bilan-gratuit/page.tsx`
+- `app/bilan-gratuit/BilanStrategiqueClient.tsx`
+- `app/bilan-gratuit/selected-offer.ts`
+- `__tests__/lib/bilan-gratuit-form.test.tsx`
+- `__tests__/marketing/echeancier-reconciliation.test.ts`
+- `e2e/pages-public-bilan-gratuit.spec.ts`
+
+## Tests executes
+
+- `npm run lint`
+- `npm run typecheck`
+- `npm run test -- --runInBand`
+- `npm run test:e2e`
+- `npm run test -- --runInBand __tests__/lib/bilan-gratuit-form.test.tsx __tests__/marketing/echeancier-reconciliation.test.ts`
+- `npm run test:e2e -- e2e/pages-public-bilan-gratuit.spec.ts`
+- `npm run build`
+- `npm run check:docs-archive`
+- `git diff --check`
+- Smoke SSR standalone `HOSTNAME=localhost PORT=3010 node .next/standalone/server.js`
+
+## Resultats
+
+- Le test e2e RED a d'abord echoue sur l'absence d'offre reperee dans le HTML initial.
+- Apres correction, le spec `/bilan-gratuit` passe avec 8 tests verts.
+- L'HTML initial de `/bilan-gratuit?offer=term-spe-simple` contient l'offre reperee, ClicToPay et Banque Zitouna, sans RIB/IBAN, et sans `BAILOUT_TO_CLIENT_SIDE_RENDERING`.
+- Gate local complet vert : Jest `497 passed, 1 skipped` suites / `6278 passed, 4 skipped` tests ; Playwright public `186 passed` ; build Next `139/139` pages.
+- Smoke SSR standalone vert sur `/`, `/offres`, `/recommandation`, `/bilan-gratuit`, `/bilan-gratuit?offer=term-spe-simple`, `/stages`, `/plateforme-aria`, `/accompagnement-scolaire`, `/contact` : HTTP 200, H1 unique, aucune occurrence visible de `undefined`.
+
+## Risques restants
+
+- Carte Nexus necessite un arbitrage Shark : loss-leader assume ou correction du perimetre/prix.
+- L'unification de charte necessite une decision Shark sur le systeme cible.
+
+## Rollback
+
+- Replacer la lecture des query params dans le composant client et retirer `selected-offer.ts`, en acceptant le retour au fallback client.
